requirements.txt
mss
opencv-python
Pillow
pyarmor
pycryptodome
pyinstaller==5.12.0
PySide6==6.4.0.1
numpy==1.26.4
pywin32
requests
wmi